removing client if error happen

+++ includes/utils/UtilityMethod.hpp
@@ -18,7 +18,7 @@ class UtilityMethod
         static std::vector<std::string> stringSpliter(const std::string& line, const std::string& delim);
         static size_t  count(const std::string& line, char to_count);
         static int switchEvents(const int& _ws, uint32_t mode, struct epoll_event& event, IO& ev);
-        static std::string numberToString(const int& number);
+        static std::string numberToString(const size_t& number);
         static const std::string& getMimeType(std::string& path, const std::string& index_path, const std::string& index, const bool& to_check);
         static std::string getFileExtension(const std::string& content_type, const short int& opt);
         static std::string decimalToHex(const int& number);
